feat(apis/pricing): complete gql implementation

+++ apis/pricing/src/models/external_data_transfer_request.rs
@@ -1,10 +1,10 @@
+use async_graphql::InputObject;
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Serialize, Deserialize, Clone)]
-#[serde(rename_all = "kebab-case")]
+#[derive(InputObject, Debug, Serialize, Deserialize, Clone)]
 pub struct ExternalDataTransferRequest {
     pub from_region_code: Option<String>,
-    pub start_range: Option<i32>,
+    pub start_range: Option<i64>,
     pub sort_by: Option<String>,
     pub sort_order: Option<String>,
 }
